Picture-Editor-App/requirements.txt
streamlit
pillow
numpy
matplotlib
seaborn
